build: make Deno resolve declarations for workspace-linked packages
Deno follows the workspace symlink to a plain file:// path, so it never applies
the exports map and type-checks the emitted JavaScript rather than the
declarations. That is why deno test surfaced errors like "Expected 2 arguments,
but got 1" against dist/cesr/parse.js, where an optional parameter is invisible.

postprocess-dist.ts (renamed from rewrite-declarations.ts) now also adds a
@ts-self-types pragma pointing each emitted file at its own .d.ts. deno check
passes, so the consumer-deno job drops --no-check and type-checks for real.

Consumers installing the published tarballs resolve declarations through the
types condition and ignore the pragma; verified both paths.
